requirements.txt
formation-studio
requests
pypresence
pillow
pyperclip